Added PingingThresholdsConfigValidatorTest, both validators now tested. Moved all tests into their own, proper namespaces. Cleaned up some comments.
