--- PingBoard.Tests/PingingTests/PingingConfigurationTests/PingingBehaviorConfigValidatorTest.cs
@@ -1,92 +1,0 @@
-using PingBoard.Pinging.Configuration;
-using PingBoard.Pinging;
-using Limits = PingBoard.Pinging.Configuration.PingingBehaviorConfigLimits;
-using FluentValidation;
-using FluentValidation.TestHelper;
-using FluentValidation.Results;
-
-
-public class PingingBehaviorConfigValidatorTest{
-
-   
-    /// <summary>
-    /// Can slice this string to generate a string of a specified length, useful for faking the PayloadStr
-    /// </summary>
-    private readonly string PayloadStrSource = new String('@', Limits.LongestAllowedPayloadStr+1);
-
-    [Fact]
-    public void BehaviorIsValidWhenAllPropertiesAtUpperLimits(){
-        PingingBehaviorConfig testConfig = new PingingBehaviorConfig{
-            PayloadStr   = PayloadStrSource.Substring(0,Limits.LongestAllowedPayloadStr),
-            PingsPerCall = Limits.MostAllowedPingsPerCall,
-            TimeoutMs    = Limits.LongestAllowedTimeoutMs,
-            Ttl          = Limits.LongestAllowedTtl,
-            WaitMs       = Limits.LongestAllowedWaitMs
-        };
-
-        PingingBehaviorConfigLimits behaviorLimits = new PingingBehaviorConfigLimits(); 
-        PingingBehaviorConfigValidator behaviorValidator = new PingingBehaviorConfigValidator(behaviorLimits);
-        TestValidationResult<PingingBehaviorConfig> result = behaviorValidator.TestValidate(testConfig);
-        result.ShouldNotHaveAnyValidationErrors();
-    }
-
-    [Fact]
-    public void BehaviorIsValidWhenAllPropertiesAtlowerLimits(){
-        PingingBehaviorConfig testConfig = new PingingBehaviorConfig{
-            PayloadStr   = PayloadStrSource.Substring(0,Limits.ShortestAllowedPayloadStr),
-            PingsPerCall = Limits.FewestAllowedPingsPerCall,
-            TimeoutMs    = Limits.ShortestAllowedTimeoutMs,
-            Ttl          = Limits.ShortestAllowedTtl,
-            WaitMs       = Limits.ShortestAllowedWaitMs
-        };
-
-        PingingBehaviorConfigLimits behaviorLimits = new PingingBehaviorConfigLimits(); 
-        PingingBehaviorConfigValidator behaviorValidator = new PingingBehaviorConfigValidator(behaviorLimits);
-        TestValidationResult<PingingBehaviorConfig> result = behaviorValidator.TestValidate(testConfig);
-        result.ShouldNotHaveAnyValidationErrors();
-    }
-
-    [Fact]
-    public void BehaviorIsInvalidWhenAllPropertiesBelowLimits(){
-        PingingBehaviorConfig testConfig = new PingingBehaviorConfig{
-            PayloadStr   = "",
-            PingsPerCall = Limits.FewestAllowedPingsPerCall-1,
-            TimeoutMs    = Limits.ShortestAllowedTimeoutMs-1,
-            Ttl          = Limits.ShortestAllowedTtl-1,
-            WaitMs       = Limits.ShortestAllowedWaitMs-1
-        };
-
-        PingingBehaviorConfigLimits behaviorLimits = new PingingBehaviorConfigLimits(); 
-        PingingBehaviorConfigValidator behaviorValidator = new PingingBehaviorConfigValidator(behaviorLimits);
-        TestValidationResult<PingingBehaviorConfig> result = behaviorValidator.TestValidate(testConfig);
-        result.ShouldHaveValidationErrorFor(test => test.PayloadStr);
-        result.ShouldHaveValidationErrorFor(test => test.PingsPerCall);
-        result.ShouldHaveValidationErrorFor(test => test.TimeoutMs);
-        result.ShouldHaveValidationErrorFor(test => test.Ttl);
-        result.ShouldHaveValidationErrorFor(test => test.WaitMs);    
-    }
-
-    [Fact]
-    public void BehaviorIsInvalidWhenAllPropertiesAboveLimits(){
-        PingingBehaviorConfig testConfig = new PingingBehaviorConfig{
-            PayloadStr   = PayloadStrSource,
-            PingsPerCall = Limits.MostAllowedPingsPerCall+1,
-            TimeoutMs    = Limits.LongestAllowedTimeoutMs+1,
-            Ttl          = Limits.LongestAllowedTtl+1,
-            WaitMs       = Limits.LongestAllowedWaitMs+1
-        };
-
-        PingingBehaviorConfigLimits behaviorLimits = new PingingBehaviorConfigLimits(); 
-        PingingBehaviorConfigValidator behaviorValidator = new PingingBehaviorConfigValidator(behaviorLimits);
-        TestValidationResult<PingingBehaviorConfig> result = behaviorValidator.TestValidate(testConfig);
-        result.ShouldHaveValidationErrorFor(test => test.PayloadStr);
-        result.ShouldHaveValidationErrorFor(test => test.PingsPerCall);
-        result.ShouldHaveValidationErrorFor(test => test.TimeoutMs);
-        result.ShouldHaveValidationErrorFor(test => test.Ttl);
-        result.ShouldHaveValidationErrorFor(test => test.WaitMs);    
-    }
-
-
-
-
-}
